Update: app_core.jsにrollup.jsを使うように
