+++ package.json
@@ -24,6 +24,7 @@
     "node-sass": "^4.9.3",
     "pug": "^2.0.3",
     "rollup": "^0.64.1",
+    "rollup-plugin-coffee-script": "^2.0.0",
     "rollup-plugin-typescript2": "^0.16.1",
     "rollup-stream": "^1.24.1",
     "sharp": "^0.20.5",
