+++ src/core/URL.ts
@@ -1,502 +1,497 @@
-///<reference path="HTTP.ts" />
-///<reference path="../app.ts" />
 ///<reference path="../global.d.ts" />
+import * as HTTP from "./HTTP"
 
-namespace app {
-  export namespace URL {
-    export const CH_BOARD_REG = /^(https?:\/\/[\w\.]+\/(?:\w+\/)?test\/(?:read\.cgi|-)\/\w+\/\d+).*$/;
-    export const CH_BOARD_REG2 = /^(https?:\/\/[\w\.]+\/\w+)\/?(?!test)$/;
-    export const CH_BOARD_ULA_REG = /^(https?):\/\/ula\.5ch\.net\/2ch\/(\w+)\/([\w\.]+)\/(\d+).*$/;
-    export const MACHI_BOARD_REG = /^(https?:\/\/\w+\.machi\.to\/bbs\/read\.cgi\/\w+\/\d+).*$/;
-    export const SHITARABA_BOARD_REG = /^(https?):\/\/jbbs\.(?:livedoor\.jp|shitaraba\.net)\/(bbs\/read(?:_archive)?\.cgi\/\w+\/\d+\/\d+).*$/;
-    export const SHITARABA_ARCHIVE_REG = /^(https?):\/\/jbbs\.(?:livedoor\.jp|shitaraba\.net)\/(\w+\/\d+)\/storage\/(\d+)\.html$/;
-    export const CH_THREAD_REG = /^(https?:\/\/[\w\.]+\/(?:subback\/|test\/-\/)?\w+\/)(?:#.*)?$/;
-    export const SHITARABA_THREAD_REG = /^(https?):\/\/jbbs\.(?:livedoor\.jp|shitaraba\.net)\/(\w+\/\d+\/)(?:#.*)?$/;
-    export function fix (url:string):string {
-      return (
-        url
-          //2ch.net->5ch.net
-          .replace(/^(https?):\/\/(\w+)\.2ch\.net\//, "$1://$2.5ch.net/")
-          // スレ系 誤爆する事は考えられないので、パラメータ部分をバッサリ切ってしまう
-          .replace(CH_BOARD_REG, "$1/")
-          .replace(CH_BOARD_REG2, "$1/")
-          .replace(CH_BOARD_ULA_REG, "$1://$3/test/read.cgi/$2/$4/")
-          .replace(MACHI_BOARD_REG, "$1/")
-          .replace(SHITARABA_BOARD_REG, "$1://jbbs.shitaraba.net/$2/")
-          .replace(SHITARABA_ARCHIVE_REG, "$1://jbbs.shitaraba.net/bbs/read_archive.cgi/$2/$3/")
-          // 板系 完全に誤爆を少しでも減らすために、パラメータ形式も限定する
-          .replace(CH_THREAD_REG, "$1")
-          .replace(SHITARABA_THREAD_REG, "$1://jbbs.shitaraba.net/$2")
-      );
-    }
+export const CH_BOARD_REG = /^(https?:\/\/[\w\.]+\/(?:\w+\/)?test\/(?:read\.cgi|-)\/\w+\/\d+).*$/;
+export const CH_BOARD_REG2 = /^(https?:\/\/[\w\.]+\/\w+)\/?(?!test)$/;
+export const CH_BOARD_ULA_REG = /^(https?):\/\/ula\.5ch\.net\/2ch\/(\w+)\/([\w\.]+)\/(\d+).*$/;
+export const MACHI_BOARD_REG = /^(https?:\/\/\w+\.machi\.to\/bbs\/read\.cgi\/\w+\/\d+).*$/;
+export const SHITARABA_BOARD_REG = /^(https?):\/\/jbbs\.(?:livedoor\.jp|shitaraba\.net)\/(bbs\/read(?:_archive)?\.cgi\/\w+\/\d+\/\d+).*$/;
+export const SHITARABA_ARCHIVE_REG = /^(https?):\/\/jbbs\.(?:livedoor\.jp|shitaraba\.net)\/(\w+\/\d+)\/storage\/(\d+)\.html$/;
+export const CH_THREAD_REG = /^(https?:\/\/[\w\.]+\/(?:subback\/|test\/-\/)?\w+\/)(?:#.*)?$/;
+export const SHITARABA_THREAD_REG = /^(https?):\/\/jbbs\.(?:livedoor\.jp|shitaraba\.net)\/(\w+\/\d+\/)(?:#.*)?$/;
+export function fix (url:string):string {
+  return (
+    url
+      //2ch.net->5ch.net
+      .replace(/^(https?):\/\/(\w+)\.2ch\.net\//, "$1://$2.5ch.net/")
+      // スレ系 誤爆する事は考えられないので、パラメータ部分をバッサリ切ってしまう
+      .replace(CH_BOARD_REG, "$1/")
+      .replace(CH_BOARD_REG2, "$1/")
+      .replace(CH_BOARD_ULA_REG, "$1://$3/test/read.cgi/$2/$4/")
+      .replace(MACHI_BOARD_REG, "$1/")
+      .replace(SHITARABA_BOARD_REG, "$1://jbbs.shitaraba.net/$2/")
+      .replace(SHITARABA_ARCHIVE_REG, "$1://jbbs.shitaraba.net/bbs/read_archive.cgi/$2/$3/")
+      // 板系 完全に誤爆を少しでも減らすために、パラメータ形式も限定する
+      .replace(CH_THREAD_REG, "$1")
+      .replace(SHITARABA_THREAD_REG, "$1://jbbs.shitaraba.net/$2")
+  );
+}
 
-    export interface GuessResult {
-      type: string;
-      bbsType: string;
-    }
-    export function guessType (url:string):GuessResult {
-      url = fix(url);
+export interface GuessResult {
+  type: string;
+  bbsType: string;
+}
+export function guessType (url:string):GuessResult {
+  url = fix(url);
 
-      if (/^https?:\/\/jbbs\.shitaraba\.net\/bbs\/read(?:_archive)?\.cgi\/\w+\/\d+\/\d+\/$/.test(url)) {
-        return {type: "thread", bbsType: "jbbs"};
-      }
-      else if (/^https?:\/\/jbbs\.shitaraba\.net\/\w+\/\d+\/$/.test(url)) {
-        return {type: "board", bbsType: "jbbs"};
-      }
-      else if (/^https?:\/\/\w+\.machi\.to\/bbs\/read\.cgi\/\w+\/\d+\/$/.test(url)) {
-        return {type: "thread", bbsType: "machi"};
-      }
-      else if (/^https?:\/\/\w+\.machi\.to\/\w+\/$/.test(url)) {
-        return {type: "board", bbsType: "machi"};
-      }
-      else if (/^https?:\/\/[\w\.]+\/(?:\w+\/)?test\/(?:read\.cgi|-)\/\w+\/\d+\/$/.test(url)) {
-        return {type: "thread", bbsType: "2ch"};
-      }
-      else if (/^https?:\/\/(?:find|info|p2|ninja)\.5ch\.net\/\w+\/$/.test(url)) {
-        return {type: "unknown", bbsType: "unknown"};
-      }
-      else if (/^https?:\/\/\w+\.(?:[25]ch|open2ch|bbspink)\.\w+\/(?:subback\/|test\/-\/)?\w+\/?$/.test(url)) {
-        return {type: "board", bbsType: "2ch"};
-      }
-      else if (/^https?:\/\/(\w+\.?)\w+\.\w+\/(?:subback\/|test\/-\/)?\w+\/?$/.test(url)) {
-        return {type: "board", bbsType: "2ch"};
-      }
-      else {
-        return {type: "unknown", bbsType: "unknown"};
-      }
-    }
+  if (/^https?:\/\/jbbs\.shitaraba\.net\/bbs\/read(?:_archive)?\.cgi\/\w+\/\d+\/\d+\/$/.test(url)) {
+    return {type: "thread", bbsType: "jbbs"};
+  }
+  else if (/^https?:\/\/jbbs\.shitaraba\.net\/\w+\/\d+\/$/.test(url)) {
+    return {type: "board", bbsType: "jbbs"};
+  }
+  else if (/^https?:\/\/\w+\.machi\.to\/bbs\/read\.cgi\/\w+\/\d+\/$/.test(url)) {
+    return {type: "thread", bbsType: "machi"};
+  }
+  else if (/^https?:\/\/\w+\.machi\.to\/\w+\/$/.test(url)) {
+    return {type: "board", bbsType: "machi"};
+  }
+  else if (/^https?:\/\/[\w\.]+\/(?:\w+\/)?test\/(?:read\.cgi|-)\/\w+\/\d+\/$/.test(url)) {
+    return {type: "thread", bbsType: "2ch"};
+  }
+  else if (/^https?:\/\/(?:find|info|p2|ninja)\.5ch\.net\/\w+\/$/.test(url)) {
+    return {type: "unknown", bbsType: "unknown"};
+  }
+  else if (/^https?:\/\/\w+\.(?:[25]ch|open2ch|bbspink)\.\w+\/(?:subback\/|test\/-\/)?\w+\/?$/.test(url)) {
+    return {type: "board", bbsType: "2ch"};
+  }
+  else if (/^https?:\/\/(\w+\.?)\w+\.\w+\/(?:subback\/|test\/-\/)?\w+\/?$/.test(url)) {
+    return {type: "board", bbsType: "2ch"};
+  }
+  else {
+    return {type: "unknown", bbsType: "unknown"};
+  }
+}
 
-    export const TSLD_REG = /^https?:\/\/(?:\w+\.)*(\w+\.\w+)\//;
-    export function tsld (url:string):string {
-      var res:any;
+export const TSLD_REG = /^https?:\/\/(?:\w+\.)*(\w+\.\w+)\//;
+export function tsld (url:string):string {
+  var res:any;
 
-      res = TSLD_REG.exec(url);
-      return res ? res[1] : "";
-    }
+  res = TSLD_REG.exec(url);
+  return res ? res[1] : "";
+}
 
-    export function getDomain (urlstr: string):string {
-      var start = urlstr.indexOf("://")+3;
-      return urlstr.slice(start, urlstr.indexOf("/", start));
-    }
+export function getDomain (urlstr: string):string {
+  var start = urlstr.indexOf("://")+3;
+  return urlstr.slice(start, urlstr.indexOf("/", start));
+}
 
-    export function getScheme (urlstr: string): string {
-      return urlstr.slice(0, urlstr.indexOf("://"));
-    }
+export function getScheme (urlstr: string): string {
+  return urlstr.slice(0, urlstr.indexOf("://"));
+}
 
-    export function setScheme (urlstr: string, protocol: string): string {
-      var split;
+export function setScheme (urlstr: string, protocol: string): string {
+  var split;
 
-      split = urlstr.indexOf("://");
-      return protocol + "://" + urlstr.slice(split+3);
-    }
+  split = urlstr.indexOf("://");
+  return protocol + "://" + urlstr.slice(split+3);
+}
 
-    export function changeScheme (urlstr: string): string {
-      var split, protocol;
+export function changeScheme (urlstr: string): string {
+  var split, protocol;
 
-      split = urlstr.indexOf("://")
-      protocol = (urlstr.slice(0, split) == "http") ? "https" : "http";
+  split = urlstr.indexOf("://")
+  protocol = (urlstr.slice(0, split) == "http") ? "https" : "http";
 
-      return protocol + "://" + urlstr.slice(split+3);
-    }
+  return protocol + "://" + urlstr.slice(split+3);
+}
 
-    export function getResNumber (urlstr: string): string|null {
-      var tmp: string[]|null;
+export function getResNumber (urlstr: string): string|null {
+  var tmp: string[]|null;
 
-      tmp = /^https?:\/\/[\w\.]+\/(?:\w+\/)?test\/(?:read\.cgi|-)\/\w+\/\d+\/(?:i|g\?g=)?(\d+).*$/.exec(urlstr);
-      if (tmp !== null) {
-        return tmp[1];
-      }
-      tmp = /^https?:\/\/ula\.5ch\.net\/2ch\/\w+\/[\w\.]+\/\d+\/(\d+).*$/.exec(urlstr);
-      if (tmp !== null) {
-        return tmp[1];
-      }
-      tmp = /^https?:\/\/\w+\.machi\.to\/bbs\/read\.cgi\/\w+\/\d+\/(\d+).*$/.exec(urlstr);
-      if (tmp !== null) {
-        return tmp[1];
-      }
-      tmp = /^https?:\/\/jbbs\.(?:livedoor\.jp|shitaraba\.net)\/bbs\/read(?:_archive)?\.cgi\/\w+\/\d+\/\d+\/(\d+).*$/.exec(urlstr);
-      if (tmp !== null) {
-        return tmp[1];
-      }
+  tmp = /^https?:\/\/[\w\.]+\/(?:\w+\/)?test\/(?:read\.cgi|-)\/\w+\/\d+\/(?:i|g\?g=)?(\d+).*$/.exec(urlstr);
+  if (tmp !== null) {
+    return tmp[1];
+  }
+  tmp = /^https?:\/\/ula\.5ch\.net\/2ch\/\w+\/[\w\.]+\/\d+\/(\d+).*$/.exec(urlstr);
+  if (tmp !== null) {
+    return tmp[1];
+  }
+  tmp = /^https?:\/\/\w+\.machi\.to\/bbs\/read\.cgi\/\w+\/\d+\/(\d+).*$/.exec(urlstr);
+  if (tmp !== null) {
+    return tmp[1];
+  }
+  tmp = /^https?:\/\/jbbs\.(?:livedoor\.jp|shitaraba\.net)\/bbs\/read(?:_archive)?\.cgi\/\w+\/\d+\/\d+\/(\d+).*$/.exec(urlstr);
+  if (tmp !== null) {
+    return tmp[1];
+  }
 
-      return null;
-    }
+  return null;
+}
 
-    export function threadToBoard (url:string):string {
-      return (
-        fix(url)
-          .replace(/^(https?):\/\/([\w\.]+)\/(?:test|bbs)\/read\.cgi\/(\w+)\/\d+\/$/, "$1://$2/$3/")
-          .replace(/^(https?):\/\/jbbs\.shitaraba\.net\/bbs\/read(?:_archive)?\.cgi\/(\w+)\/(\d+)\/\d+\/$/, "$1://jbbs.shitaraba.net/$2/$3/")
-      );
-    }
+export function threadToBoard (url:string):string {
+  return (
+    fix(url)
+      .replace(/^(https?):\/\/([\w\.]+)\/(?:test|bbs)\/read\.cgi\/(\w+)\/\d+\/$/, "$1://$2/$3/")
+      .replace(/^(https?):\/\/jbbs\.shitaraba\.net\/bbs\/read(?:_archive)?\.cgi\/(\w+)\/(\d+)\/\d+\/$/, "$1://jbbs.shitaraba.net/$2/$3/")
+  );
+}
 
-    export function parseQuery (urlStr:string, fromSearch:boolean = true):{[index:string]:any;} {
-      if (fromSearch) {
-        return new URLSearchParams(urlStr.slice(1));
-      }
-      return (new window.URL(urlStr)).searchParams;
-    }
+export function parseQuery (urlStr:string, fromSearch:boolean = true):{[index:string]:any;} {
+  if (fromSearch) {
+    return new URLSearchParams(urlStr.slice(1));
+  }
+  return (new window.URL(urlStr)).searchParams;
+}
 
-    export function parseHashQuery (url:string):{[index:string]:any;} {
-      var tmp;
+export function parseHashQuery (url:string):{[index:string]:any;} {
+  var tmp;
 
-      tmp = /#(.+)$/.exec(url);
+  tmp = /#(.+)$/.exec(url);
 
-      return tmp ? new URLSearchParams(tmp[1]) : new URLSearchParams();
-    }
+  return tmp ? new URLSearchParams(tmp[1]) : new URLSearchParams();
+}
 
-    export function buildQuery (data:{[index:string]:any;}) {
-      return (new URLSearchParams(<any>data)).toString();
-    }
+export function buildQuery (data:{[index:string]:any;}) {
+  return (new URLSearchParams(<any>data)).toString();
+}
 
-    export const SHORT_URL_LIST = new Set([
-      "amba.to",
-      "amzn.to",
-      "bit.ly",
-      "buff.ly",
-      "cas.st",
-      "cos.lv",
-      "dlvr.it",
-      "ekaz10.xyz",
-      "fb.me",
-      "g.co",
-      "goo.gl",
-      "htn.to",
-      "ift.tt",
-      "is.gd",
-      "itun.es",
-      "j.mp",
-      "jump.cx",
-      "kkbox.fm",
-      "morimo2.info",
-      "ow.ly",
-      "p.tl",
-      "prt.nu",
-      "redd.it",
-      "snipurl.com",
-      "spoti.fi",
-      "t.co",
-      "tiny.cc",
-      "tinyurl.com",
-      "tl.gd",
-      "tr.im",
-      "trib.al",
-      "qq4q.biz",
-      "u0u1.net",
-      "ur0.biz",
-      "ur0.work",
-      "url.ie",
-      "urx.nu",
-      "urx.red",
-      "urx2.nu",
-      "urx3.nu",
-      "ur0.pw",
-      "ur2.link",
-      "ustre.am",
-      "ux.nu",
-      "wb2.biz",
-      "wk.tk",
-      "xrl.us"
-    ]);
+export const SHORT_URL_LIST = new Set([
+  "amba.to",
+  "amzn.to",
+  "bit.ly",
+  "buff.ly",
+  "cas.st",
+  "cos.lv",
+  "dlvr.it",
+  "ekaz10.xyz",
+  "fb.me",
+  "g.co",
+  "goo.gl",
+  "htn.to",
+  "ift.tt",
+  "is.gd",
+  "itun.es",
+  "j.mp",
+  "jump.cx",
+  "kkbox.fm",
+  "morimo2.info",
+  "ow.ly",
+  "p.tl",
+  "prt.nu",
+  "redd.it",
+  "snipurl.com",
+  "spoti.fi",
+  "t.co",
+  "tiny.cc",
+  "tinyurl.com",
+  "tl.gd",
+  "tr.im",
+  "trib.al",
+  "qq4q.biz",
+  "u0u1.net",
+  "ur0.biz",
+  "ur0.work",
+  "url.ie",
+  "urx.nu",
+  "urx.red",
+  "urx2.nu",
+  "urx3.nu",
+  "ur0.pw",
+  "ur2.link",
+  "ustre.am",
+  "ux.nu",
+  "wb2.biz",
+  "wk.tk",
+  "xrl.us"
+]);
 
-    export async function expandShortURL (shortUrl: string): Promise<string> {
-      var res, finalUrl = "";
-      var cache = new app.Cache(shortUrl);
+export async function expandShortURL (shortUrl: string): Promise<string> {
+  var res, finalUrl = "";
+  var cache = new app.Cache(shortUrl);
 
-      res = await ( async () => {
-        try {
-          await cache.get();
-          return {data: cache.data, url: null};
-        } catch {
-          var req = new app.HTTP.Request("HEAD", shortUrl);
-          req.timeout = parseInt(app.config.get("expand_short_url_timeout")!);
+  res = await ( async () => {
+    try {
+      await cache.get();
+      return {data: cache.data, url: null};
+    } catch {
+      var req = new HTTP.Request("HEAD", shortUrl);
+      req.timeout = parseInt(app.config.get("expand_short_url_timeout")!);
 
-          var {status, responseURL: resUrl} = await req.send();
-
-          if (status >= 400) {
-            return {data: null, url: null};
-          }
-          // 無限ループの防止
-          if (resUrl === shortUrl) {
-            return {data: null, url: null};
-          }
-
-          // 取得したURLが短縮URLだった場合は再帰呼出しする
-          if (SHORT_URL_LIST.has(getDomain(resUrl))) {
-            resUrl = await expandShortURL(resUrl)
-            return {data: null, url: resUrl};
-          // 短縮URL以外なら終了
-          } else {
-            return {data: null, url: resUrl};
-          }
-        }
-      })();
-
-      if (res.data === null && res.url !== null) {
-        cache.lastUpdated = Date.now();
-        cache.data = res.url;
-        cache.put();
-        finalUrl = res.url;
-      } else if (res.data !== null && res.url === null) {
-        finalUrl = res.data;
-      }
-      return finalUrl;
-    }
-
-    const AUDIO_REG = /\.(?:mp3|m4a|wav|oga|spx)(?:[\?#:&].*)?$/;
-    const VIDEO_REG = /\.(?:mp4|m4v|webm|ogv)(?:[\?#:&].*)?$/;
-    const OGG_REG = /\.(?:ogg|ogx)(?:[\?#:&].*)?$/;
-    export function getExtType (filename: string, {
-        audio = true,
-        video = true,
-        oggIsAudio = false,
-        oggIsVideo = true
-      }: Partial<{
-        audio: boolean,
-        video: boolean,
-        oggIsAudio: boolean,
-        oggIsVideo: boolean
-      }> = {}
-    ): string|null {
-      if (audio && AUDIO_REG.test(filename)) {
-        return "audio";
-      }
-      if (video && VIDEO_REG.test(filename)) {
-        return "video";
-      }
-      if (video && oggIsVideo && OGG_REG.test(filename)) {
-        return "video";
-      }
-      if (audio && oggIsAudio && OGG_REG.test(filename)) {
-        return "audio";
-      }
-      return null;
-    }
-
-    export function convertUrlFromPhone (url: string): string {
-      var regs: any[];
-      var tmp: string[]|null = [];
-      var mode: string;
-      var scheme: string = "";
-      var server: string|null = null;
-      var board: string|null = null;
-      var thread: string|null = null;
-      var resUrl: string = url;
-
-      function checkReg (value: any, index: number, array: any[]): boolean {
-        return (tmp = value.exec(url));
-      }
-
-      mode = tsld(url);
-      switch (mode) {
-        case "2ch.net":
-        case "5ch.net":
-          regs = [
-            /(https?):\/\/itest\.5ch\.net\/(?:\w+\/)?test\/read\.cgi\/(\w+)\/(\d+)\//,
-            /(https?):\/\/itest\.5ch\.net\/(?:subback\/)?(\w+)(?:\/)?/,
-            /(https?):\/\/c\.2ch\.net\/test\/-\/(\w+)\/(\d+)\//,
-            /(https?):\/\/c\.2ch\.net\/test\/-\/(\w+)\//
-          ]
-          if (regs.some(checkReg)) {
-            scheme = tmp[1];
-            board = tmp[2];
-            thread = tmp[3] ? tmp[3] : null;
-            if (board !== null) {
-              if (serverNet.has(board) === true) {
-                server = serverNet.get(board)!;
-              // 携帯用bbspinkの可能性をチェック
-              } else if (serverPink.has(board) === true) {
-                server = serverPink.get(board)!;
-                mode = "bbspink.com";
-              }
-            }
-          }
-          break;
-
-        case "2ch.sc":
-          regs = [
-            /(https?):\/\/sp\.2ch\.sc\/(?:\w+\/)?test\/read\.cgi\/(\w+)\/(\d+)\//,
-            /(https?):\/\/sp\.2ch\.sc\/(?:subback\/)?(\w+)\//
-          ]
-          if (regs.some(checkReg)) {
-            scheme = tmp[1];
-            board = tmp[2];
-            thread = tmp[3] ? tmp[3] : null;
-            if (board !== null && serverSc.has(board) === true) {
-              server = serverSc.get(board)!;
-            }
-          }
-          break;
-
-        case "bbspink.com":
-          regs = [
-            /(https?):\/\/itest\.bbspink\.com\/(?:\w+\/)?test\/read\.cgi\/(\w+)\/(\d+)\//,
-            /(https?):\/\/itest\.bbspink\.com\/(?:subback\/)?(\w+)(?:\/)?/
-          ]
-          if (regs.some(checkReg)) {
-            scheme = tmp[1];
-            board = tmp[2];
-            thread = tmp[3] ? tmp[3] : null;
-            if (board !== null && serverPink.has(board) === true) {
-              server = serverPink.get(board)!;
-            }
-          }
-          break;
-      }
-
-      if (server !== null) {
-        if (thread !== null) {
-          resUrl = `${scheme}://${server}.${mode}/test/read.cgi/${board}/${thread}/`;
-        } else {
-          resUrl = `${scheme}://${server}.${mode}/${board}/`;
-        }
-      }
-
-      return resUrl;
-    }
-
-    var serverNet = new Map<string, string>();
-    var serverSc = new Map<string, string>();
-    var serverPink = new Map<string, string>();
-
-    function bbsmenuParam (url: string) {
-      var res = {net: false, sc: false, bbspink: false};
-      var tmp = /http:\/\/kita\.jikkyo\.org\/cbm\/cbm\.cgi\/([\w\d\.]+)(?:\/-all|-live2324)?(?:\/=[\d\w=!&$()[\]{}]+)?\/bbsmenuk?2?\.html/.exec(url);
-      if (tmp) {
-        var param = tmp[1].split(".");
-        for (var mode of param) {
-          switch (mode) {
-            case "20":
-            case "2r":
-            case "5r":
-              res.net = true; break;
-            case "sc":
-              res.sc = true; break;
-            case "p0":
-            case "p1":
-            case "p5":
-              res.bbspink = true; break;
-          }
-          if (res.net && res.sc && res.bbspink) {
-            break;
-          }
-        }
-      } else if (
-        url.endsWith("menu.2ch.net/bbsmenu.html") ||
-        url.endsWith("menu.5ch.net/bbsmenu.html")
-      ) {
-        res.net = true;
-        res.bbspink = true;
-      } else if (
-        url.endsWith("www.2ch.sc/bbsmenu.html")
-      ) {
-        res.sc = true;
-      }
-      return res;
-    }
-
-    function applyServerInfo (res: {net: boolean, sc: boolean, bbspink: boolean}, menu: any[]) {
-      var boardNet = new Map<string, string>(),
-        boardSc = new Map<string, string>(),
-        boardPink = new Map<string, string>(),
-        tmp: string[]|null;
-
-      for (let category of menu) {
-        for (let board of category.board) {
-          if (res.net && (tmp = /https?:\/\/(\w+)\.5ch\.net\/(\w+)\/.*?/.exec(board.url)) !== null) {
-            boardNet.set(tmp[2], tmp[1]);
-          }else if (res.sc && (tmp = /https?:\/\/(\w+)\.2ch\.sc\/(\w+)\/.*?/.exec(board.url)) !== null) {
-            boardSc.set(tmp[2], tmp[1]);
-          }else if (res.bbspink && (tmp = /https?:\/\/(\w+)\.bbspink\.com\/(\w+)\/.*?/.exec(board.url)) !== null) {
-            boardPink.set(tmp[2], tmp[1]);
-          }
-        }
-      }
-
-      if (boardNet.size > 0) serverNet = boardNet;
-      if (boardSc.size > 0) serverSc = boardSc;
-      if (boardPink.size > 0) serverPink = boardPink;
-    }
-
-    export async function pushServerInfo (url: string, menu: any[][]): Promise<void> {
-      var res = bbsmenuParam(url);
-
-      applyServerInfo(res, menu);
-
-      if (res.net && res.sc && res.bbspink) {
-        return;
-      }
-
-      var param = "";
-      if (!res.net) param += "20.";
-      if (!res.sc) param += "sc.";
-      if (!res.bbspink) param += "p0.";
-      param += "99";
-      var url = `http://kita.jikkyo.org/cbm/cbm.cgi/${param}/-all/bbsmenu.html`;
-      var menu = <any[][]>(await app.BBSMenu.fetch(url, false)).menu
-      var res = bbsmenuParam(url);
-      applyServerInfo(res, menu);
-    }
-
-    function exchangeNetSc (url: string): string|null {
-      var mode: string[]|null;
-      var server: string|null = null;
-      var target: string|null = null;
-      var resUrl: string|null = null;
-
-      mode = /(https?):\/\/(\w+)\.(5ch\.net|2ch\.sc)\/test\/read\.cgi\/(\w+)\/(\d+)\//.exec(url);
-      if (mode === null) {
-        return null;
-      }
-
-      if (mode[3] === "5ch.net") {
-        if (serverSc.has(mode[4]) === true) {
-          server = serverSc.get(mode[4])!;
-          target = "2ch.sc";
-        }
-      } else {
-        if (serverNet.has(mode[4]) === true) {
-          server = serverNet.get(mode[4])!;
-          target = "5ch.net";
-        }
-      }
-
-      if (server !== null) {
-        resUrl = `${mode[1]}://${server}.${target}/test/read.cgi/${mode[4]}/${mode[5]}/`;
-      }
-      return resUrl;
-    }
-
-    export async function convertNetSc (url: string): Promise<string> {
-      var newUrl = exchangeNetSc(url);
-
-      if (newUrl) {
-        return newUrl;
-      }
-      var scheme: string;
-      var tmpUrl: string;
-      var tmp: string[]|null;
-
-      tmp = /(https?):\/\/(\w+)\.5ch\.net\/test\/read\.cgi\/(\w+\/\d+\/)/.exec(url);
-      if (tmp === null) {
-        throw new Error("不明なURL形式です");
-      }
-      tmpUrl = `http://${tmp[2]}.2ch.sc/test/read.cgi/${tmp[3]}`;
-      scheme = tmp[1];
-
-      var req = new app.HTTP.Request("HEAD", tmpUrl);
       var {status, responseURL: resUrl} = await req.send();
+
       if (status >= 400) {
-        throw new Error("移動先情報の取得の通信に失敗しました");
+        return {data: null, url: null};
       }
-      tmp = /https?:\/\/(\w+)\.2ch\.sc\/test\/read\.cgi\/(\w+)\/(\d+)\//.exec(resUrl);
-      if (tmp !== null) {
-        if (serverSc.has(tmp[2]) === false) {
-          serverSc.set(tmp[2], tmp[1]);
+      // 無限ループの防止
+      if (resUrl === shortUrl) {
+        return {data: null, url: null};
+      }
+
+      // 取得したURLが短縮URLだった場合は再帰呼出しする
+      if (SHORT_URL_LIST.has(getDomain(resUrl))) {
+        resUrl = await expandShortURL(resUrl)
+        return {data: null, url: resUrl};
+      // 短縮URL以外なら終了
+      } else {
+        return {data: null, url: resUrl};
+      }
+    }
+  })();
+
+  if (res.data === null && res.url !== null) {
+    cache.lastUpdated = Date.now();
+    cache.data = res.url;
+    cache.put();
+    finalUrl = res.url;
+  } else if (res.data !== null && res.url === null) {
+    finalUrl = res.data;
+  }
+  return finalUrl;
+}
+
+const AUDIO_REG = /\.(?:mp3|m4a|wav|oga|spx)(?:[\?#:&].*)?$/;
+const VIDEO_REG = /\.(?:mp4|m4v|webm|ogv)(?:[\?#:&].*)?$/;
+const OGG_REG = /\.(?:ogg|ogx)(?:[\?#:&].*)?$/;
+export function getExtType (filename: string, {
+    audio = true,
+    video = true,
+    oggIsAudio = false,
+    oggIsVideo = true
+  }: Partial<{
+    audio: boolean,
+    video: boolean,
+    oggIsAudio: boolean,
+    oggIsVideo: boolean
+  }> = {}
+): string|null {
+  if (audio && AUDIO_REG.test(filename)) {
+    return "audio";
+  }
+  if (video && VIDEO_REG.test(filename)) {
+    return "video";
+  }
+  if (video && oggIsVideo && OGG_REG.test(filename)) {
+    return "video";
+  }
+  if (audio && oggIsAudio && OGG_REG.test(filename)) {
+    return "audio";
+  }
+  return null;
+}
+
+export function convertUrlFromPhone (url: string): string {
+  var regs: any[];
+  var tmp: string[]|null = [];
+  var mode: string;
+  var scheme: string = "";
+  var server: string|null = null;
+  var board: string|null = null;
+  var thread: string|null = null;
+  var resUrl: string = url;
+
+  function checkReg (value: any, index: number, array: any[]): boolean {
+    return (tmp = value.exec(url));
+  }
+
+  mode = tsld(url);
+  switch (mode) {
+    case "2ch.net":
+    case "5ch.net":
+      regs = [
+        /(https?):\/\/itest\.5ch\.net\/(?:\w+\/)?test\/read\.cgi\/(\w+)\/(\d+)\//,
+        /(https?):\/\/itest\.5ch\.net\/(?:subback\/)?(\w+)(?:\/)?/,
+        /(https?):\/\/c\.2ch\.net\/test\/-\/(\w+)\/(\d+)\//,
+        /(https?):\/\/c\.2ch\.net\/test\/-\/(\w+)\//
+      ]
+      if (regs.some(checkReg)) {
+        scheme = tmp[1];
+        board = tmp[2];
+        thread = tmp[3] ? tmp[3] : null;
+        if (board !== null) {
+          if (serverNet.has(board) === true) {
+            server = serverNet.get(board)!;
+          // 携帯用bbspinkの可能性をチェック
+          } else if (serverPink.has(board) === true) {
+            server = serverPink.get(board)!;
+            mode = "bbspink.com";
+          }
         }
-        resUrl = `${scheme}://${tmp[1]}.2ch.sc/test/read.cgi/${tmp[2]}/${tmp[3]}/`;
       }
-      return resUrl;
+      break;
+
+    case "2ch.sc":
+      regs = [
+        /(https?):\/\/sp\.2ch\.sc\/(?:\w+\/)?test\/read\.cgi\/(\w+)\/(\d+)\//,
+        /(https?):\/\/sp\.2ch\.sc\/(?:subback\/)?(\w+)\//
+      ]
+      if (regs.some(checkReg)) {
+        scheme = tmp[1];
+        board = tmp[2];
+        thread = tmp[3] ? tmp[3] : null;
+        if (board !== null && serverSc.has(board) === true) {
+          server = serverSc.get(board)!;
+        }
+      }
+      break;
+
+    case "bbspink.com":
+      regs = [
+        /(https?):\/\/itest\.bbspink\.com\/(?:\w+\/)?test\/read\.cgi\/(\w+)\/(\d+)\//,
+        /(https?):\/\/itest\.bbspink\.com\/(?:subback\/)?(\w+)(?:\/)?/
+      ]
+      if (regs.some(checkReg)) {
+        scheme = tmp[1];
+        board = tmp[2];
+        thread = tmp[3] ? tmp[3] : null;
+        if (board !== null && serverPink.has(board) === true) {
+          server = serverPink.get(board)!;
+        }
+      }
+      break;
+  }
+
+  if (server !== null) {
+    if (thread !== null) {
+      resUrl = `${scheme}://${server}.${mode}/test/read.cgi/${board}/${thread}/`;
+    } else {
+      resUrl = `${scheme}://${server}.${mode}/${board}/`;
     }
   }
+
+  return resUrl;
+}
+
+var serverNet = new Map<string, string>();
+var serverSc = new Map<string, string>();
+var serverPink = new Map<string, string>();
+
+function bbsmenuParam (url: string) {
+  var res = {net: false, sc: false, bbspink: false};
+  var tmp = /http:\/\/kita\.jikkyo\.org\/cbm\/cbm\.cgi\/([\w\d\.]+)(?:\/-all|-live2324)?(?:\/=[\d\w=!&$()[\]{}]+)?\/bbsmenuk?2?\.html/.exec(url);
+  if (tmp) {
+    var param = tmp[1].split(".");
+    for (var mode of param) {
+      switch (mode) {
+        case "20":
+        case "2r":
+        case "5r":
+          res.net = true; break;
+        case "sc":
+          res.sc = true; break;
+        case "p0":
+        case "p1":
+        case "p5":
+          res.bbspink = true; break;
+      }
+      if (res.net && res.sc && res.bbspink) {
+        break;
+      }
+    }
+  } else if (
+    url.endsWith("menu.2ch.net/bbsmenu.html") ||
+    url.endsWith("menu.5ch.net/bbsmenu.html")
+  ) {
+    res.net = true;
+    res.bbspink = true;
+  } else if (
+    url.endsWith("www.2ch.sc/bbsmenu.html")
+  ) {
+    res.sc = true;
+  }
+  return res;
+}
+
+function applyServerInfo (res: {net: boolean, sc: boolean, bbspink: boolean}, menu: any[]) {
+  var boardNet = new Map<string, string>(),
+    boardSc = new Map<string, string>(),
+    boardPink = new Map<string, string>(),
+    tmp: string[]|null;
+
+  for (let category of menu) {
+    for (let board of category.board) {
+      if (res.net && (tmp = /https?:\/\/(\w+)\.5ch\.net\/(\w+)\/.*?/.exec(board.url)) !== null) {
+        boardNet.set(tmp[2], tmp[1]);
+      }else if (res.sc && (tmp = /https?:\/\/(\w+)\.2ch\.sc\/(\w+)\/.*?/.exec(board.url)) !== null) {
+        boardSc.set(tmp[2], tmp[1]);
+      }else if (res.bbspink && (tmp = /https?:\/\/(\w+)\.bbspink\.com\/(\w+)\/.*?/.exec(board.url)) !== null) {
+        boardPink.set(tmp[2], tmp[1]);
+      }
+    }
+  }
+
+  if (boardNet.size > 0) serverNet = boardNet;
+  if (boardSc.size > 0) serverSc = boardSc;
+  if (boardPink.size > 0) serverPink = boardPink;
+}
+
+export async function pushServerInfo (url: string, menu: any[][]): Promise<void> {
+  var res = bbsmenuParam(url);
+
+  applyServerInfo(res, menu);
+
+  if (res.net && res.sc && res.bbspink) {
+    return;
+  }
+
+  var param = "";
+  if (!res.net) param += "20.";
+  if (!res.sc) param += "sc.";
+  if (!res.bbspink) param += "p0.";
+  param += "99";
+  var url = `http://kita.jikkyo.org/cbm/cbm.cgi/${param}/-all/bbsmenu.html`;
+  var menu = <any[][]>(await app.BBSMenu.fetch(url, false)).menu
+  var res = bbsmenuParam(url);
+  applyServerInfo(res, menu);
+}
+
+function exchangeNetSc (url: string): string|null {
+  var mode: string[]|null;
+  var server: string|null = null;
+  var target: string|null = null;
+  var resUrl: string|null = null;
+
+  mode = /(https?):\/\/(\w+)\.(5ch\.net|2ch\.sc)\/test\/read\.cgi\/(\w+)\/(\d+)\//.exec(url);
+  if (mode === null) {
+    return null;
+  }
+
+  if (mode[3] === "5ch.net") {
+    if (serverSc.has(mode[4]) === true) {
+      server = serverSc.get(mode[4])!;
+      target = "2ch.sc";
+    }
+  } else {
+    if (serverNet.has(mode[4]) === true) {
+      server = serverNet.get(mode[4])!;
+      target = "5ch.net";
+    }
+  }
+
+  if (server !== null) {
+    resUrl = `${mode[1]}://${server}.${target}/test/read.cgi/${mode[4]}/${mode[5]}/`;
+  }
+  return resUrl;
+}
+
+export async function convertNetSc (url: string): Promise<string> {
+  var newUrl = exchangeNetSc(url);
+
+  if (newUrl) {
+    return newUrl;
+  }
+  var scheme: string;
+  var tmpUrl: string;
+  var tmp: string[]|null;
+
+  tmp = /(https?):\/\/(\w+)\.5ch\.net\/test\/read\.cgi\/(\w+\/\d+\/)/.exec(url);
+  if (tmp === null) {
+    throw new Error("不明なURL形式です");
+  }
+  tmpUrl = `http://${tmp[2]}.2ch.sc/test/read.cgi/${tmp[3]}`;
+  scheme = tmp[1];
+
+  var req = new HTTP.Request("HEAD", tmpUrl);
+  var {status, responseURL: resUrl} = await req.send();
+  if (status >= 400) {
+    throw new Error("移動先情報の取得の通信に失敗しました");
+  }
+  tmp = /https?:\/\/(\w+)\.2ch\.sc\/test\/read\.cgi\/(\w+)\/(\d+)\//.exec(resUrl);
+  if (tmp !== null) {
+    if (serverSc.has(tmp[2]) === false) {
+      serverSc.set(tmp[2], tmp[1]);
+    }
+    resUrl = `${scheme}://${tmp[1]}.2ch.sc/test/read.cgi/${tmp[2]}/${tmp[3]}/`;
+  }
+  return resUrl;
 }
