+++ src/global.d.ts
@@ -11,12 +11,16 @@ declare namespace UI {
   var Animate: any;
 }
 declare namespace app {
-  var Cache: any;
-  var BBSMenu: any;
   var config: any;
   var Callbacks: any;
   var log: any;
   var deepCopy: any;
   var message: any;
   var defer: any;
+
+  var Cache: any;
+  var BBSMenu: any;
+  var ReadState: any;
+
+  var HTTP: any;
 }
